Build process scripting.
The main purpose of this is to convert .tmLanguage.yaml files to
.tmLanguage.json files, and to generate token patterns from the
.tmLanguage.json files.

That's done by using the webpack-shell-plugin-next to run th
a program at the start of the build. The program is
scripts/tmlanguage_yaml_to_json.py, which converts the
.tmLanguage.yaml files to .tmLanguage.json files, and then runs the
syntax_to_token_patterns.py  script to generate the token patterns.

This introduces a dependency on uv to ensure that the Python dependencies
are installed before the scripts are run.